Create separate files for `useStreamRuleTypes` and `useStreams` types.
