+++ graylog2-web-interface/src/components/streams/StreamsOverview/StreamsOverview.tsx
@@ -16,14 +16,13 @@
  */
 import PropTypes from 'prop-types';
 import React, { useState, useEffect, useCallback, useMemo } from 'react';
-import { useQuery } from '@tanstack/react-query';
 import styled from 'styled-components';
 
 import { Alert, Label } from 'components/bootstrap';
 import { Icon, IfPermitted, PaginatedList, SearchForm } from 'components/common';
 import Spinner from 'components/common/Spinner';
 import QueryHelper from 'components/common/QueryHelper';
-import type { Stream, StreamRuleType } from 'stores/streams/StreamsStore';
+import type { Stream } from 'stores/streams/StreamsStore';
 import StreamsStore from 'stores/streams/StreamsStore';
 import { StreamRulesStore } from 'stores/streams/StreamRulesStore';
 import usePaginationQueryParameter from 'hooks/usePaginationQueryParameter';
@@ -33,11 +32,12 @@ import StreamActions from 'components/streams/StreamsOverview/StreamActions';
 import { Link } from 'components/common/router';
 import Routes from 'routing/Routes';
 import type { ColumnRenderers } from 'components/common/EntityDataTable';
-import UserNotification from 'util/UserNotification';
 import IndexSetCell from 'components/streams/StreamsOverview/IndexSetCell';
 import BulkActions from 'components/streams/StreamsOverview/BulkActions';
 import ThroughputCell from 'components/streams/StreamsOverview/ThroughputCell';
 import type { SearchParams, Sort } from 'stores/PaginationTypes';
+import useStreams from 'components/streams/hooks/useStreams';
+import useStreamRuleTypes from 'components/streams/hooks/useStreaRuleTypes';
 
 import StatusCell from './StatusCell';
 
@@ -83,46 +83,6 @@ const customColumnRenderers = (indexSets: Array<IndexSet>): ColumnRenderers<Stre
   },
 });
 
-const usePaginatedStreams = (searchParams: SearchParams): { data: { streams: Array<Stream>, pagination: { total: number } } | undefined, refetch: () => void } => {
-  const { data, refetch } = useQuery(
-    ['streams', 'overview', searchParams],
-    () => StreamsStore.searchPaginated(
-      searchParams.page,
-      searchParams.pageSize,
-      searchParams.query,
-      { sort: searchParams?.sort.attributeId, direction: searchParams?.sort.direction },
-    ),
-    {
-      onError: (errorThrown) => {
-        UserNotification.error(`Loading streams failed with status: ${errorThrown}`,
-          'Could not load streams');
-      },
-      keepPreviousData: true,
-    },
-  );
-
-  return ({
-    data,
-    refetch,
-  });
-};
-
-const useStreamRuleTypes = (): { data: Array<StreamRuleType> | undefined } => {
-  const { data } = useQuery(
-    ['streams', 'rule-types'],
-    () => StreamRulesStore.types(),
-    {
-      onError: (errorThrown) => {
-        UserNotification.error(`Loading stream rule types failed with status: ${errorThrown}`,
-          'Could not load stream rule types');
-      },
-      keepPreviousData: true,
-    },
-  );
-
-  return ({ data });
-};
-
 type Props = {
   onStreamCreate: (stream: Stream) => Promise<void>,
   indexSets: Array<IndexSet>
@@ -141,7 +101,7 @@ const StreamsOverview = ({ onStreamCreate, indexSets }: Props) => {
     },
   });
   const { data: streamRuleTypes } = useStreamRuleTypes();
-  const { data: paginatedStreams, refetch: refetchStreams } = usePaginatedStreams(searchParams);
+  const { data: paginatedStreams, refetch: refetchStreams } = useStreams(searchParams);
   const columnRenderers = useMemo(() => customColumnRenderers(indexSets), [indexSets]);
 
   useEffect(() => {
@@ -193,7 +153,7 @@ const StreamsOverview = ({ onStreamCreate, indexSets }: Props) => {
   );
 
   if (!paginatedStreams || !streamRuleTypes) {
-    return (<Spinner />);
+    return <Spinner />;
   }
 
   const { streams, pagination: { total } } = paginatedStreams;
